requirements.txt
google.generativeai
requests
pandas
dotenv
streamlit
google-cloud-firestore